wasm にするには lib.rs が必要
